feat: классы элементов меню вынесены в папку core
