+++ bot/src/ui/__init__.py
@@ -1,1 +1,2 @@
-from .main_menu import main_menu
+from ui.core import MenuNames
+from ui.main_menu import main_menu
