+++ bot/src/ui/main_menu.py
@@ -1,26 +1,6 @@
 from telegram import ReplyKeyboardMarkup, Update
 from telegram.ext import CallbackContext
-from ui import constants as const
-
-
-class MenuElement:
-    def __init__(self, name: str, result: str, answers: dict[str, str]) -> None:
-        self.name = name
-        self.result = result
-        self.answers = answers
-
-    def __call__(self, answer_name: str) -> str:
-        return self.answers[answer_name]
-
-
-class MenuNames:
-    """Текстовое описание для элементов меню."""
-
-    start_menu = MenuElement(
-        name=const.NAME,
-        result=const.GREETING_MESSAGE,
-        answers=const.ANSWERS,
-    )
+from ui import MenuNames
 
 
 def main_menu(update: Update, context: CallbackContext) -> None:
